requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssselect2==0.7.0
Django==5.1.8
django-celery-beat==2.7.0
django-crontab==0.7.1
django-extensions==3.2.3
django-simple-history==3.7.0
django-tenants==3.7.0
django-timezone-field==7.0
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.55.7
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
num2words==0.5.13
openpyxl==3.1.5
packaging==25.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydot==3.0.2
pydyf==0.11.0
pyparsing==3.2.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.4.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
